logout을 UserHandler로 이동
